### TODO   - 플레이어 한테 시간주는 로직 구현   - OnAccpetProc , 네트워크 송/수신 일어나는 모든 부분에서 플레이어 시간 넣어줘야함.   - 그거한다음, client 붙여서 접속되고 서로 섹터가 잘 적용되는지 확인 (눈으로, 손으로 )    - 더미 붙여서 집에서 테스트해보기

+++ MMOFIGHTERS/Player.h
@@ -120,6 +120,6 @@ namespace Core
 		short _Y;
 
 		char _Hp;
-		DWORD _LastTime;
+		unsigned long _LastTime;
 	};
 }
